Completed functionality of adding multiple businesses against user
