+++ src/containers/admin/addNewAdmin.js
@@ -224,7 +224,7 @@ class AddAdmin extends Component {
                                     onChange={this.onImageChange.bind(this)} 
                                     accept=".jpg"
                                     />
-
+                                    <span style={{fontSize:'8pt', color:'grey'}}>File must be less than 2 MB</span>
                             </div>            
                         </div>
                         <div className="row end-xs">
