+++ src/containers/merchant/manageMerchants.js
@@ -2,33 +2,27 @@
 import React, { Component } from 'react';
 import { Field, reduxForm } from 'redux-form'
 import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import NumberFormat from "react-number-format";
 
 //material-ui
 import Paper from '@material-ui/core/Paper';
 import MenuItem from '@material-ui/core/MenuItem';
 import FormControl from '@material-ui/core/FormControl';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import TableFooter from '@material-ui/core/TableFooter';
-import TablePagination from '@material-ui/core/TablePagination';
 import Button from '@material-ui/core/Button';
-import FormLabel from '@material-ui/core/FormLabel';
 import AddIcon from '@material-ui/icons/Add';
+import Dialog from '@material-ui/core/Dialog';
+import DialogContent from '@material-ui/core/DialogContent';
 
 //Components
-import InputField from '../../components/inputField';
-import TablePaginationActions from '../../components/tableGrid';
+import UserList from '../../containers/user/userList'
 import {renderSelectField} from '../../components/selectControl';
-import DialogBox from '../../components/alertDialog'
-import Loader from '../../components/loader'
-
-//Actions
-import { getMerchantListWithFilter, deleteMerchant } from '../../actions/merchantAction';
+import InputField from '../../components/inputField';
+import Loader from '../../components/loader';
+import Register from '../../containers/account/register';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import IconButton from '@material-ui/core/IconButton';
+import CloseIcon from '@material-ui/icons/Close';
+import Slide from '@material-ui/core/Slide';
 
 //Data
 import Data from '../../staticData';
@@ -41,55 +35,22 @@ const styles = {
       selectControl:{
         fontSize: '12px',
       },
+      appBar: {
+        position: 'relative',
+      },
 };
+
+function Transition(props) {
+    return <Slide direction="up" {...props} />;
+  }
 
 class ManageMerchants extends Component {
 
     state = {
         name:'',
         status: '',
-        location:'',
-        merchantList:undefined,
-        page: 0,
-        rowsPerPage: 5,
-        dialogOpen: false,
-        permissionDisplayBox: false,
-        disableReset: true,
-        merchantId:'',
-        message:'',
+        openSignupPopUp: false,
     };
-
-    componentWillMount()
-    {
-        this.getAllMerchants();
-    }
-
-    componentWillReceiveProps(nextProps) {
-
-        if (nextProps) {
-          if (nextProps.merchantPayload){
-            if(nextProps.merchantPayload.status === 200){
-                this.setState({showLoader:false})
-                this.setState({merchantList: nextProps.merchantPayload.responseData})
-            }
-            else{
-                this.setState({showLoader:false})
-            }
-          }
-        
-          if(nextProps.merchantDelete){
-            if(nextProps.merchantDelete.status === 200){
-                this.setState({showLoader:false})
-                this.setState({ dialogOpen: true });
-                this.setState({message:nextProps.merchantDelete.message});
-                this.getAllMerchants();
-            }
-            else{                
-                this.setState({showLoader:false})
-            }
-          }
-        }
-    }
 
     //Method to handle change event for dropdown
     handleChange = event => {
@@ -103,91 +64,27 @@ class ManageMerchants extends Component {
         }
     };
 
-    handleChangePage = (event, page) => {
-        this.setState({ page });
-    };
-    
-    handleChangeRowsPerPage = event => {
-        this.setState({ rowsPerPage: event.target.value });
-    };
-
-    getAllMerchants(){
-        if(this.props.userData.user.responseData.token){
-            this.setState({showLoader:true})
-            this.props.getMerchantListWithFilter(this.state.name, this.state.status, this.state.location, this.props.userData.user.responseData.token)
-        }
-        else{
-            //#TODO : Handle token expire case
-        }
-    }
-
     onHandleSearch(){
-        this.getAllMerchants();
+        this.child.searchHandler();
     }
 
-
-    deleteMerchant = (merchantId, inactive) => {
-        if (this.state.permissionDisplayBox) {
-            this.handleClose();
-            if(this.props.userData.user.responseData.token){
-                this.setState({showLoader:true})
-                this.props.deleteMerchant(this.state.merchantId, this.state.inactive, this.props.userData.user.responseData.token);
-            }
-            else{
-                //#TODO: Handle token expire case here
-            }
-        }
-        else{
-            this.setState({ permissionDisplayBox: true, merchantId: merchantId, inactive : inactive });
-        }
-        
-    }
-
-    updateMerchant = (merchantId) => {
-        this.props.history.push({pathname:'/updateMerchant',state: { detail: merchantId }})
+    addNewMerchant(){
+        this.setState({ openSignupPopUp: true });
     }
 
     handleClose = () => {
-        this.setState({ dialogOpen: false });
-        this.setState({ permissionDisplayBox: false });
+        this.setState({ openSignupPopUp: false });
     };
-
-    addNewMerchant(){
-        this.props.history.push('/addNewMerchant')
-    }
 
     onHandleReset(){
         this.setState({name:''});
         this.setState({status:''});
-        this.setState({location:''});
         this.setState({disableReset:true});
         this.props.reset();
-
-        if(this.props.userData.user.responseData.token){
-            this.props.getMerchantListWithFilter("", "", "", this.props.userData.user.responseData.token)
-        }
-    
+        this.child.resetHandler();
     }
 
     render() {
-
-        const { merchantList, rowsPerPage, page, dialogOpen, permissionDisplayBox, inactive } = this.state;
-        const emptyRows = rowsPerPage - Math.min(rowsPerPage, (merchantList !== undefined ? merchantList.length : 0) - page * rowsPerPage);
-
-        const actions = [
-            <Button key="ok" onClick={this.handleClose} color="primary" autoFocus>
-                OK
-            </Button>
-        ];
-
-        const permissionActions = [
-            <Button key="no" onClick={this.handleClose} color="primary">
-                No
-            </Button>,
-            <Button key="yes" onClick={this.deleteMerchant} color="primary" autoFocus>
-                Yes
-            </Button>,
-        ];
 
         return (
           <div className="row">
@@ -195,17 +92,29 @@ class ManageMerchants extends Component {
             <Loader status={this.state.showLoader} />
 
             <div>
-                <DialogBox 
-                    displayDialogBox={dialogOpen} 
-                    message={this.state.message} 
-                    actions={actions} 
-                />
-                <DialogBox 
-                    displayDialogBox={permissionDisplayBox} 
-                    message=  {inactive === true ? "Are you sure to deactivate merchant account?" : "Are you sure to activate merchant account?"}
-                    actions={permissionActions} 
-                />
-            </div> 
+                <Dialog
+                    fullScreen
+                    open={this.state.openSignupPopUp}
+                    onClose={this.handleClose.bind(this)}
+                    fullWidth={true}
+                    maxWidth = {'sm'}
+                    scroll = 'paper'
+                    TransitionComponent={Transition}
+                >
+                    <AppBar style={styles.appBar} backgroundColor="#5CBBFF">
+                        <Toolbar>
+                        <IconButton color="inherit" onClick={this.handleClose} aria-label="Close">
+                            <CloseIcon />
+                        </IconButton>
+                        </Toolbar>
+                    </AppBar>
+                    <div>
+                        <DialogContent>
+                            <Register history={this.props.history} /> 
+                        </DialogContent>
+                    </div>
+                </Dialog>
+            </div>
 
             <div className="row">
             <div className="col-xs-12">
@@ -238,7 +147,7 @@ class ManageMerchants extends Component {
                                         Status
                                     </MenuItem>
                                     {
-                                    Data.searchStatus.map((item) =>{
+                                    Data.userStatus.map((item) =>{
                                         return <MenuItem 
                                             style={styles.selectControl}
                                             key={item.id}
@@ -250,31 +159,6 @@ class ManageMerchants extends Component {
                                 </Field>    
                             </FormControl>  
                         </div>
-                        <div className="col-xs-12 col-sm-6 col-md-2">
-                            <FormControl style={styles.formControl}>
-                                <Field
-                                    name="location"
-                                    component={renderSelectField}
-                                    fullWidth={true}
-                                    onChange={this.handleChange}
-                                    displayEmpty
-                                    >
-                                    <MenuItem value="" disabled>
-                                        Location
-                                    </MenuItem>
-                                    {
-                                    Data.states.map((item) =>{
-                                        return <MenuItem 
-                                            style={styles.selectControl}
-                                            key={item.id}
-                                            value={item.prefix}>
-                                            {item.name}
-                                        </MenuItem>
-                                    })
-                                    }
-                                </Field>    
-                            </FormControl>  
-                        </div>    
                         <div className="col-xs-12 col-sm-6 col-md-3">
                             <button 
                                 type="button"
@@ -291,7 +175,7 @@ class ManageMerchants extends Component {
                                 > Filter
                             </button> 
                         </div>       
-                        <div className="col-xs-12 col-sm-6 col-md-3 end-md">
+                        <div className="col-xs-12 col-sm-6 col-md-5 end-md">
                             <Button 
                                 variant="fab"
                                 type="button"
@@ -308,118 +192,22 @@ class ManageMerchants extends Component {
             </div>
             </div>
 
-            <div className="row">
-            <div className="col-xs-12">
-                    <Paper className="pagePaper">
-                    <div className="tableWrapperMaterial">
-                    <Table className="tableMaterial">
-                        <TableHead>
-                            <TableRow>
-                                <TableCell numeric>#</TableCell>
-                                <TableCell>Merchant Name</TableCell>
-                                <TableCell>Location</TableCell>
-                                <TableCell>Email Address</TableCell>
-                                <TableCell>Phone Number</TableCell>
-                                <TableCell>Status</TableCell>
-                                <TableCell>Actions</TableCell>
-                            </TableRow>
-                        </TableHead>
-                        <TableBody>
-                        { 
-                            (merchantList !== undefined) ? (
-                            (merchantList.length === 0) ? 
-                                (
-                                    <TableRow style={{ height: 48 * emptyRows }}>
-                                        <TableCell colSpan={7}>
-                                            <div className="dashboardText" style={{textAlign:"center", width:"100%"}} ><b>No Record Found</b></div>
-                                        </TableCell>
-                                    </TableRow>                                
-                                )
-                                : (
-                                merchantList
-                                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                                .map((object, index) => {
-                                    return (
-                                    <TableRow className="tableRow" key={object.id}>
-                                        <TableCell numeric>{object.serial_number}</TableCell>
-                                        <TableCell>{object.first_v + " " + object.last_v}</TableCell>
-                                        <TableCell>{object.city_v}</TableCell>
-                                        <TableCell>{object.email_v}</TableCell>
-                                        <TableCell><NumberFormat value={object.phone_v} displayType={'text'} format="+1 (###) ###-####" /></TableCell>
-                                        <TableCell>
-                                            <div className={object.inactive_v === 1 ? "titleRed" : "titleGreen"}><FormLabel component="label" style={{color:'white', fontSize:'12px'}}>{object.status}</FormLabel></div>
-                                        </TableCell>
-                                        <TableCell> 
-                                            <div className="row start-xs" style={{marginRight:'0px',marginBottom:'0px'}}>
-                                                <div className="col-xs-6">
-                                                    <button 
-                                                        type="button" 
-                                                        disabled={object.inactive_v === 1 ? true : false} 
-                                                        onClick={() => this.updateMerchant(object.id)} 
-                                                        className={object.inactive_v === 1 ? "disabledButton" : "enabledButton"}
-                                                        style={ ((index%2 !== 0) ? {backgroundColor:'#ffffff', height: '100%'} : {backgroundColor:'#f2f6f2', height:'100%'})}
-                                                        > 
-                                                        <img src="../images/ic_edit.svg" alt="" /> 
-                                                    </button>
-                                                </div>
-                                                <div className="col-xs-6">
-                                                    <button 
-                                                        type="button" 
-                                                        onClick={() => this.deleteMerchant(object.id, !object.inactive_v)} 
-                                                        className="enabledButton"
-                                                        style={ ((index%2 !== 0) ? {backgroundColor:'#ffffff', height: '100%'} : {backgroundColor:'#f2f6f2', height:'100%'})}
-                                                        > 
-                                                        {
-                                                            object.inactive_v === 1 ?
-                                                                <img src="../images/switch_off.svg" alt="" />
-                                                            :
-                                                                <img src="../images/switch_on.svg" alt="" />
-                                                        }
-                                                    </button>
-                                                </div>
-                                            </div>
-                                        </TableCell>    
-                                    </TableRow>
-                                    );
-                                })
-                                )
-                                ):(
-                                    <TableRow style={{ height: 48 * emptyRows }}>
-                                        <TableCell colSpan={7}>
-                                            <div className="dashboardText" style={{textAlign:"center", width:"100%"}} ><b>No Record Available</b></div>
-                                        </TableCell>
-                                    </TableRow>
-                                )
-                            }
-                        </TableBody>
-                        <TableFooter>
-                            <TableRow>
-                                <TablePagination
-                                colSpan={7}
-                                count={(merchantList !== undefined) ? merchantList.length : 0}
-                                rowsPerPage={rowsPerPage}
-                                page={page}
-                                onChangePage={this.handleChangePage}
-                                onChangeRowsPerPage={this.handleChangeRowsPerPage}
-                                ActionsComponent={TablePaginationActions}
-                                />
-                            </TableRow>
-                        </TableFooter>
-                    </Table>
-                    </div>
-                </Paper>
-              </div>
-            </div>   
+            <UserList 
+                name={this.state.name}
+                status={this.state.status}
+                userType="Merchant"
+                resetUserType={false}
+                history={this.props.history}
+                isHover={true}
+                isClick={true}
+                onRef={ref => (this.child = ref)} 
+            />
+            
         </div> 
         </div>
         );
     }
 }
-
-
-const mapDispatchToProps = (dispatch) => {
-    return bindActionCreators({ getMerchantListWithFilter, deleteMerchant }, dispatch)
-  }
   
   ManageMerchants = connect(
     state => ({
@@ -427,7 +215,6 @@ const mapDispatchToProps = (dispatch) => {
       merchantPayload: state.merchant.merchantList === undefined ? undefined : state.merchant.merchantList,
       merchantDelete: state.merchant.deleteMerchant === undefined ? undefined : state.merchant.deleteMerchant
     }),
-    mapDispatchToProps,
   )(ManageMerchants)
   
   export default reduxForm({form: 'FrmManageMerchant'})(ManageMerchants)
